end of section 10 backend

+++ Application/Activities/Delete.cs
@@ -1,4 +1,5 @@
-﻿using MediatR;
+﻿using Application.Core;
+using MediatR;
 using Persistence;
 using System;
 using System.Collections.Generic;
@@ -8,14 +9,14 @@ using System.Threading.Tasks;
 
 namespace Application.Activities
 {
-    public class Delete 
+    public class Delete
     {
-        public class Command : IRequest
+        public class Command : IRequest<Result<Unit>>
         {
             public Guid Id { get; set; }
         }
 
-        public class Handler : IRequestHandler<Command>
+        public class Handler : IRequestHandler<Command, Result<Unit>>
         {
             private readonly DataContext context;
 
@@ -24,13 +25,25 @@ namespace Application.Activities
                 this.context = context;
             }
 
-            public async Task Handle(Command request, CancellationToken cancellationToken)
+            public async Task<Result<Unit>> Handle(Command request, CancellationToken cancellationToken)
             {
                 var activity = await context.Activities.FindAsync(request.Id, cancellationToken);
-                
+
+                if (activity == null)
+                {
+                    return null;
+                }
+
                 context.Activities.Remove(activity);
 
-                await context.SaveChangesAsync(cancellationToken);
+                var result = await context.SaveChangesAsync(cancellationToken) > 0;
+
+                if (!result)
+                {
+                    return Result<Unit>.Failure("Failed to delete activity");
+                }
+
+                return Result<Unit>.Success(Unit.Value);
             }
         }
     }
